+++ gulpfile.js
@@ -61,6 +61,7 @@ gulp.task('js-merge', function () {
     './src/78show.js',
     './src/79set.js',
     './src/80console.js',
+    './src/81commit.js',
     './src/85help.js',
     './src/87source.js',
     './src/89assert.js',