+++ test/future/test147.js
@@ -5,16 +5,11 @@ if(typeof exports === 'object') {
 	__dirname = '.';
 };
 
-
-if(typeof exports != 'object') {
-
 describe('Test 147 - WebSQL Database', function() {
 
-
-if(false){
 	it("1. Nested SQL", function(done){
-		var res = alasql('CREATE WebSQL DATABASE wd147 ("1.0","My database",1024*1024) AS test147', [], function(res){
-			console.log('done', res);
+		var res = alasql('CREATE WebSQL DATABASE wd147 ("1.0","My database",1024*1024) AS test147');
+		console.log('done', res);
 			done();
 //			alasql('ATTACH WebSQL DATABASE wd147', [], function() {
 //				alasql('USE wd147',[],function() {
@@ -22,15 +17,11 @@ if(false){
 //				});
 //			});
 		// });
-		});
 	});
 
 	it("99. Detach database", function(done){
 		 // Do we really need this?
 		done();
 	});
-}
-
 });
 
-};