+++ test/test148.js
@@ -12,7 +12,7 @@ describe('Test 148 - EXPLAIN SELECT', function() {
 		alasql('CREATE TABLE one (a INT)');
 		alasql('INSERT INTO one VALUES (1),(2),(3)');
 		var res = alasql('EXPLAIN SELECT * FROM one WHERE a IN (SELECT * FROM one) ORDER BY a');
-		console.table(res);
+//		console.table(res);
 		done();
 	});