+++ test/test150.js
@@ -13,6 +13,8 @@ describe('Test 150 - localStorage Engine', function() {
 		alasql('SET AUTOCOMMIT OFF');
 		assert(!alasql.autocommit);
 
+//delete localStorage['ls150.one'];
+
 		alasql('DROP localStorage DATABASE IF EXISTS ls150');
 		assert(!localStorage['ls150']);
 		assert(!localStorage['ls150.one']);
@@ -30,7 +32,7 @@ describe('Test 150 - localStorage Engine', function() {
 	});
 
 	it("3. Attach localStorage database", function(done){
-		alasql('ATTACH localStorage DATABASE ls149 AS test150');
+		alasql('ATTACH localStorage DATABASE ls150 AS test150');
 		assert(alasql.databases.test150);
 		assert(alasql.databases.test150.engineid == 'localStorage');
 		done();
@@ -39,10 +41,10 @@ describe('Test 150 - localStorage Engine', function() {
 	it("4. Create localStorage databases", function(done){
 		alasql('CREATE TABLE IF NOT EXISTS test150.one (a int, b string)');
 //		assert(!alasql.databases.test149.tables.one);
-console.log(JSON.parse(localStorage['ls150']));
-		assert(JSON.parse(localStorage['ls150']).tables);
+//console.log(JSON.parse(localStorage['ls150']));
+		assert(localStorage['ls150.one']);
+		assert(JSON.parse(localStorage['ls150']).tables.one);
 //		assert(JSON.parse(localStorage['ls149']).tables.one);
-		assert(!localStorage['ls150.one']);
 		// var tb = JSON.parse(localStorage['ls149']).tables.one;
 		// assert(tb.columns);
 		// assert(tb.columns[0].columnid == 'a');
@@ -55,26 +57,42 @@ console.log(JSON.parse(localStorage['ls150']));
 		alasql('CREATE TABLE test150a.one (a int, b string)');
 
 		alasql('insert into test150a.one VALUES (1,"Moscow"), (2, "Kyiv"), (3,"Minsk")');
-		alasql('select * into test150.one from test150a.one');
-		assert(alasql.databases.test150.tables.one.data, [{"a":1,"b":"Moscow"},{"a":2,"b":"Kyiv"},{"a":3,"b":"Minsk"}]);
+		var res = alasql('select * into test150.one from test150a.one');
+//		console.log(res);
+		assert.deepEqual(alasql.databases.test150.tables.one.data, [{"a":1,"b":"Moscow"},{"a":2,"b":"Kyiv"},{"a":3,"b":"Minsk"}]);
 
 		var res = alasql('select * from test150.one');
-		assert(res, [{"a":1,"b":"Moscow"},{"a":2,"b":"Kyiv"},{"a":3,"b":"Minsk"}]);
+//		console.log(res);
+		assert.deepEqual(res, [{"a":1,"b":"Moscow"},{"a":2,"b":"Kyiv"},{"a":3,"b":"Minsk"}]);
 		done();
 	});
 
-/*;	it("6.Select from localStorage table", function(done) {
-		var res = alasql('SELECT * FROM test149.one');
+	it("6.Select from localStorage table", function(done) {
+		var res = alasql('SELECT * FROM test150.one');
+//		console.log(res);
 		assert(res.length == 3);
 		done();
 	});
-*/
 
 	it("7.Select into localStorage table", function(done) {
 		var res = alasql('SELECT a*2 as a, b INTO test150.one FROM test150.one');
 		assert(res == 3);
+		var res = alasql('SELECT * FROM test150.one');
+		assert(res.length == 6);
 		done();
 	});
+
+	it("8.Select into localStorage table", function(done) {
+		alasql('USE test150');
+		var res = alasql('COMMIT');
+//		console.log(res);
+		assert(res,1);
+
+		var res = alasql('SELECT * FROM test150.one');
+		assert(res.length == 6);
+		done();
+	});
+
 
 	it("8.Drop localStorage table", function(done) {
 		alasql('DROP TABLE test150.one');